feat: implement reading library detail

+++ src/main/java/com/dope/wb/repository/LibraryRepository.java
@@ -3,5 +3,5 @@ package com.dope.wb.repository;
 import com.dope.wb.domain.board.library.Library;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface LibraryRepository extends JpaRepository<Library, Integer> {
+public interface LibraryRepository extends JpaRepository<Library, Long> {
 }
