+++ src/main/java/com/dope/wb/repository/ProductSketchRepository.java
@@ -4,6 +4,6 @@ import com.dope.wb.domain.board.product.Product;
 import com.dope.wb.domain.board.attachment.ProductSketch;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProductSketchRepository extends JpaRepository<ProductSketch, Integer> {
+public interface ProductSketchRepository extends JpaRepository<ProductSketch, Long> {
     ProductSketch findByProduct(Product product);
 }
